Create xml_to_html folder and add ticha_preview script
All the HTML conversion utilities are now consolidated in the
xml_to_html folder. The ticha_magic script once again outputs a
barebones HTML document, and the ticha_preview script generates the full
HTML with CSS for previewing. This allows us to keep simple the server-side
code that paginates the HTML, while still being able to view an accurate
preview.